Functioning Baseline

+++ Models/ProfileManager.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Windows;
 using SummonerSwap.Helpers;
 using SummonerSwap.Services;
 
@@ -18,11 +19,18 @@ namespace SummonerSwap.Models
 
         public void SaveProfile(string name)
         {
+            if (!RiotClientService.CheckLeagueClientRunning())
+            {
+                MessageBox.Show("Please be logged into League of Legends to save an account!");
+                return;
+            }
+
             string profilePath = Path.Combine(ProfilesPath, name);
             if (Directory.Exists(profilePath))
                 Directory.Delete(profilePath, true);
 
             FileManager.CopyDirectory(RiotDataPath, profilePath);
+            System.Diagnostics.Debug.WriteLine($"Profile saved to {profilePath}");
         }
 
         public void LoadProfile(string name)
@@ -35,6 +43,16 @@ namespace SummonerSwap.Models
             Directory.Delete(RiotDataPath, true);
             FileManager.CopyDirectory(profilePath, RiotDataPath);
             RiotClientService.LaunchClient();
+        }
+
+        public void DeleteProfile(string name)
+        {
+            string profilePath = Path.Combine(ProfilesPath, name);
+            if (Directory.Exists(profilePath))
+            {
+                Directory.Delete(profilePath, true);
+                MessageBox.Show($"Profile '{name}' deleted successfully.");
+            }
         }
 
         public IEnumerable<string> ListProfiles()
